+++ components/layout/header/header.story.js
@@ -1,0 +1,17 @@
+import React from 'react'
+
+import LayoutHeader from './header'
+
+// More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
+export default {
+	title: 'Layout/Header',
+	component: LayoutHeader,
+	// More on argTypes: https://storybook.js.org/docs/react/api/argtypes
+}
+
+// More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
+const Template = (args) => <LayoutHeader {...args} />
+
+export const Primary = Template.bind({})
+// More on args: https://storybook.js.org/docs/react/writing-stories/args
+Primary.args = {}