+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.1-alpha.1](https://github.com/empaticDev/charonium-frontend/compare/v0.1.1-alpha.0...v0.1.1-alpha.1) (2022-03-30)
+
+
+### Features
+
+* Init storybook base ([129cd81](https://github.com/empaticDev/charonium-frontend/commit/129cd81c4408abfa78f82b6194900eafef86d168))
+
+
+### Storybook
+
+* Add Badge support ([4b639b7](https://github.com/empaticDev/charonium-frontend/commit/4b639b7423224192da796181a114d6226a76e832))
+* Add default readme to storybook as introduction ([0d67cea](https://github.com/empaticDev/charonium-frontend/commit/0d67cea09d3d5f5e55480fcb16d6ff5523daf0fb))
+* Refactor Storybook documentation ([0f36e8a](https://github.com/empaticDev/charonium-frontend/commit/0f36e8a533c8ddeeb61c20ff150c96804e108b90))
+
 ### [0.1.1-alpha.0](https://github.com/empaticDev/charonium-frontend/compare/v0.1.0...v0.1.1-alpha.0) (2022-03-29)